Disable source-map and remove js caching in webpack config
In the webpack.common.js file, the 'devtool' property was removed to disable the source-map for debugging in the browser. This change was made to improve the production build's performance and to discourage source code exposure. Also, the caching of JavaScript files was removed to avoid serving stale code, which can lead to unexpected behaviors on the client side.