Optimize webpack for production build
Added a minimizer in the optimization field of the webpack production config. This minimizer is a CssMinimizerPlugin configured to use its lightning fast CSS minify utility. The addition of this minimizer will help ensure our production builds are as lean and efficient as possible, improving load times and reducing network overhead.